requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
bidict==0.23.1; python_version >= '3.8'
blessed==1.20.0; python_version >= '2.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
fastapi==0.110.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1
idna==3.6; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
pluggy==1.4.0; python_version >= '3.8'
pydantic==2.6.3; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pytest==8.0.2; python_version >= '3.8'
python-dotenv==1.0.1
python-engineio==4.9.0; python_version >= '3.6'
python-socketio[client]==5.11.1; python_version >= '3.8'
pyyaml==6.0.1
requests==2.31.0; python_version >= '3.7'
simple-websocket==1.0.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.36.3; python_version >= '3.8'
typing-extensions==4.10.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
uvicorn[standard]==0.27.1; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wsproto==1.2.0; python_full_version >= '3.7.0'
